[TM-1303] Pull demographic processing into a trait and apply to RestorationPartner.
